+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/ProductDefinitionRepository.kt
@@ -1,7 +1,6 @@
 package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data
 
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.ProductDefinition
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.SingleMetricProductDefinition
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.SingleReportProductDefinition
 
 interface ProductDefinitionRepository {
@@ -11,6 +10,4 @@ interface ProductDefinitionRepository {
   fun getProductDefinition(definitionId: String, dataProductDefinitionsPath: String? = null): ProductDefinition
 
   fun getSingleReportProductDefinition(definitionId: String, reportId: String, dataProductDefinitionsPath: String? = null): SingleReportProductDefinition
-
-  fun getSingleMetricProductDefinition(definitionId: String, metricId: String, dataProductDefinitionsPath: String? = null): SingleMetricProductDefinition
 }