+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Below you can find the changes included in each release.
 
+## 7.0.0 
+Removed the metrics definition and metrics data endpoints.
+Updated the DPD with regard to dashboards and metrics.
+Updated the dashboard definition response.
+
 ## 6.3.2
 Add Crosstab supporting properties